+++ src/types/AdInfo.ts
@@ -1,3 +1,5 @@
+import type { ErrorCode } from '../ErrorCode';
+
 /**
  * Represents an ad that has been served by AppLovin MAX.
  */
@@ -70,7 +72,7 @@ export type AdLoadFailedInfo = {
     /**
      * The error code for the error.
      */
-    code: string | null;
+    code: ErrorCode;
 
     /**
      * The error message for the error.
@@ -105,7 +107,7 @@ export type AdDisplayFailedInfo = AdInfo & {
     /**
      * The error code for the error.
      */
-    code: string | null;
+    code: ErrorCode;
 
     /**
      * The error message for the error.
@@ -271,7 +273,7 @@ export type AdErrorInfo = {
     /**
      * The error code for the error.
      */
-    code: number;
+    code: ErrorCode;
 
     /**
      * The error message for the error.